requirements.txt
kivy==2.3.0
kivymd==1.2.0
Pillow==10.4.0
numpy==1.26.4
pylibdmtx==0.1.9
pytesseract==0.3.9
opencv-python==4.6.0.66
setuptools==72.1.0

